backend/requirements.txt
Django==5.0.6
djangorestframework==3.15.1
django-filter==24.2
django-cors-headers==4.3.1
psycopg2-binary==2.9.9
requests==2.31.0